get pcsx2 compiling on aarch64

+++ pcsx2/IPU/mpeg2lib/Vlc.h
@@ -159,7 +159,7 @@ static const __aligned16 MVtab MV_10 [] = {
 
 
 static const DMVtab DMV_2 [] = {
-    { 0, 1}, { 0, 1}, { 1, 2}, {-1, 2}
+    { 0, 1}, { 0, 1}, { 1, 2}, {(s8)-1, 2}
 };
